+++ CHANGELOG.md
@@ -4,6 +4,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.2.0] - 13.11.2022
+
+### Added
+- `UpdateClientAsync` to match `PUT /workspaces/{workspaceId}/clients/{clientId}`
+- `ArchiveAndDeleteClientAsync` to easy delete client
+- `UpdateTagAsync` to match `PUT /workspaces/{workspaceId}/tags/{tagId}`
+- `ArchiveAndDeleteTagAsync` to easy delete tag
+
+### Changed
+- Marked `UpdateClientNameAsync` as `**deprecated**
+- `ClientUpdateDto.Archived` to `bool?`
+- `DetailedReportRequest.Project` to type `ProjectFilterDto`
+
 ## [2.1.2] - 27.10.2022
 
 ### Fixed